+++ src/fhir/data-types/primitive/UnsignedIntType.ts
@@ -21,28 +21,28 @@
  *
  */
 
-import { fhirUri, fhirUriSchema } from './primitive-types';
+import { fhirUnsignedInt, fhirUnsignedIntSchema } from './primitive-types';
 import { PrimitiveType } from '@src/fhir/base-models/core-fhir-models';
 import { PrimitiveTypeError } from '@src/fhir/errors/PrimitiveTypeError';
 
 /**
- * Primitive type "uri" in FHIR: A Uniform Resource Identifier Reference (RFC 3986)
+ * Primitive type "unsignedInt" in FHIR: Any non-negative integer in the range 0..2,147,483,647
  *
- * @see {@link https://hl7.org/fhir/R5/datatypes.html#uri|uri}
+ * @see {@link https://hl7.org/fhir/R5/datatypes.html#unsignedInt|unsignedInt}
  */
-export class UriType extends PrimitiveType<fhirUri> {
-  constructor(value?: fhirUri) {
+export class UnsignedIntType extends PrimitiveType<fhirUnsignedInt> {
+  constructor(value?: fhirUnsignedInt) {
     super();
     this.setValue(value);
   }
 
-  public override setValue(value?: fhirUri): this {
+  public override setValue(value?: fhirUnsignedInt): this {
     if (value !== undefined) {
-      const parseResult = fhirUriSchema.safeParse(value);
+      const parseResult = fhirUnsignedIntSchema.safeParse(value);
       if (parseResult.success) {
         super.setValue(parseResult.data);
       } else {
-        throw new PrimitiveTypeError(`Invalid value (${value}) for UriType`, parseResult.error);
+        throw new PrimitiveTypeError(`Invalid value for UnsignedIntType`, parseResult.error);
       }
     } else {
       super.setValue(undefined);
@@ -50,35 +50,36 @@ export class UriType extends PrimitiveType<fhirUri> {
     return this;
   }
 
-  public encode(value: fhirUri): string {
-    const parseResult = fhirUriSchema.safeParse(value);
+  public encode(value: fhirUnsignedInt): string {
+    const parseResult = fhirUnsignedIntSchema.safeParse(value);
     if (parseResult.success) {
       return parseResult.data.toString();
     } else {
-      throw new PrimitiveTypeError(`Invalid value (${value}) for UriType`, parseResult.error);
+      throw new PrimitiveTypeError(`Invalid value for UnsignedIntType`, parseResult.error);
     }
   }
 
-  public parse(value: string): fhirUri {
-    const parseResult = fhirUriSchema.safeParse(value);
+  public parse(value: string): fhirUnsignedInt {
+    const valueNumber = Number.parseInt(value);
+    const parseResult = fhirUnsignedIntSchema.safeParse(valueNumber);
     if (parseResult.success) {
       return parseResult.data;
     } else {
-      throw new PrimitiveTypeError(`Invalid value (${value}) for UriType`, parseResult.error);
+      throw new PrimitiveTypeError(`Invalid value for UnsignedIntType`, parseResult.error);
     }
   }
 
   public override fhirType(): string {
-    return 'uri';
+    return 'unsignedInt';
   }
 
-  public override copy(): UriType {
-    const dest = new UriType();
+  public override copy(): UnsignedIntType {
+    const dest = new UnsignedIntType();
     this.copyValues(dest);
     return dest;
   }
 
-  public override copyValues(dest: UriType): void {
+  public override copyValues(dest: UnsignedIntType): void {
     super.copyValues(dest);
     dest.setValueAsString(this.getValueAsString());
   }